// src/app/models/maintainer.ts
export interface Maintainer {
    id: string;
    assetCategoryId: string;
    organizationId: string;
    companyName: string;
    maintainerName: string;
    telephone: string;
    officePhone: string;
    org2: string;
    categoryFirstLevel: string;
    categorySecondLevel: string;
    categoryThirdLevel: string;
}